% Generated by build script
\name{wqp_characteristics}
\alias{wqp_characteristics}
\title{List available characteristics (parameters) for a location}
\usage{
wqp_characteristics()
}
\arguments{
\item{siteid}{Monitoring location ID (e.g. "USGS-01646500")}
\item{top}{Max records (default 100)}
}
\value{tibble: characteristic, result_count}
\description{List available characteristics (parameters) for a location}
