% Generated by build script
\name{wqp_station_detail}
\alias{wqp_station_detail}
\title{Fetch station details with summary statistics}
\usage{
wqp_station_detail()
}
\arguments{
\item{siteid}{Monitoring location ID (e.g. "USGS-01646500")}
}
\value{tibble: organization, monitoring_location_id, monitoring_location_name,}
\description{Fetch station details with summary statistics}
