% Generated by build script
\name{wqp_stations}
\alias{wqp_stations}
\title{Fetch monitoring stations from the Water Quality Portal}
\usage{
wqp_stations()
}
\arguments{
\item{countycode}{County FIPS code (e.g. "US:06:037")}
\item{statecode}{State FIPS code (e.g. "US:06")}
\item{top}{Max number of records (default 100)}
}
\value{tibble: organization, monitoring_location_id, monitoring_location_name,}
\description{Fetch monitoring stations from the Water Quality Portal}
