% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wqp.R
\name{wqp_stations}
\alias{wqp_stations}
\title{Fetch monitoring stations from the Water Quality Portal}
\usage{
wqp_stations(countycode = NULL, statecode = NULL, top = 100)
}
\arguments{
\item{countycode}{County FIPS code (e.g. "US:06:037")}

\item{statecode}{State FIPS code (e.g. "US:06")}

\item{top}{Max number of records (default 100)}
}
\value{
tibble: organization, monitoring_location_id, monitoring_location_name,
location_type, latitude, longitude, state_code, county_code, huc_code
}
\description{
Fetch monitoring stations from the Water Quality Portal
}
